tec: Extract a LocaleSorter from OrganizationSorter
